Reorganização de código
